lisa/tests: Move kernel tests up one directory
