+++ lisa/exekall_customize.py
@@ -31,8 +31,8 @@ from lisa.env import TestEnv, TargetConf
 from lisa.platforms.platinfo import PlatformInfo
 from lisa.utils import HideExekallID, Loggable, ArtifactPath, get_subclasses, groupby, Serializable
 from lisa.conf import MultiSrcConf
-from lisa.tests.kernel.test_bundle import TestBundle, Result, ResultBundle, CannotCreateError
-from lisa.tests.kernel.scheduler.load_tracking import FreqInvarianceItem
+from lisa.tests.test_bundle import TestBundle, Result, ResultBundle, CannotCreateError
+from lisa.tests.scheduler.load_tracking import FreqInvarianceItem
 
 from exekall.utils import info, get_name, get_mro, NoValue
 from exekall.engine import ExprData, Consumer, PrebuiltOperator, ValueDB
@@ -369,4 +369,3 @@ RESULT_TAG_MAP = {
 }
 # Make sure we cover all cases
 assert set(RESULT_TAG_MAP.keys()) == set(Result)
-
