+++ tests/test_test_bundle.py
@@ -17,7 +17,7 @@
 
 from lisa.platforms.platinfo import PlatformInfo
 
-from lisa.tests.kernel.test_bundle import TestBundle, ResultBundle
+from lisa.tests.test_bundle import TestBundle, ResultBundle
 from .utils import create_local_testenv, StorageTestCase
 
 class DummyTestBundle(TestBundle):
@@ -45,7 +45,7 @@ class DummyTestBundle(TestBundle):
 
 class BundleCheck(StorageTestCase):
     """
-    A test class that verifies some :class:`lisa.tests.kernel.test_bundle.TestBundle`
+    A test class that verifies some :class:`lisa.tests.test_bundle.TestBundle`
     base behaviours.
     """
     def setUp(self):
